Before
```console
➜  ethabi git:(master) cargo run -- encode params -v '(bool,bool[])[]' '[(true,[false,true])]'
    Finished dev [unoptimized + debuginfo] target(s) in 0.04s
     Running `target/debug/ethabi encode params -v '(bool,bool[])[]' '[(true,[false,true])]'`
Error: Invalid data

➜  ethabi git:(master) cargo run -- encode params -v '(bool,bool[])' '(true,[false,true])' 
    Finished dev [unoptimized + debuginfo] target(s) in 0.11s
     Running `target/debug/ethabi encode params -v '(bool,bool[])' '(true,[false,true])'`
Error: Invalid data
```

After
```console
➜  ethabi git:(parse-tuples-arrays) cargo run -- encode params -v '(bool,bool[])[]' '[(true,[false,true])]'
    Finished dev [unoptimized + debuginfo] target(s) in 0.04s
     Running `target/debug/ethabi encode params -v '(bool,bool[])[]' '[(true,[false,true])]'`
00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001

➜  ethabi git:(parse-tuples-arrays) cargo run -- encode params -v '(bool,bool[])' '(true,[false,true])'
    Finished dev [unoptimized + debuginfo] target(s) in 0.04s
     Running `target/debug/ethabi encode params -v '(bool,bool[])' '(true,[false,true])'`
000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
```