@vkgnosis 2022-08-08
CI fails because of cargo-fmt. I'm not familiar with this code (and I don't think anyone of the maintainers still is) but the test case makes sense to me and the code looks fine at a glance.